requirements.txt
pypdf2
pdfplumber
chromadb
openai
python-dotenv
langchain
langchain-community
langchain-openai
tiktoken
flask
flask-cors

